fix: refactored sampling methods and added tabs to select them
